+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [Gradle Plugin] Implement stricter MigrationFile versioning (#5730 by @madisp)
 
 ### Fixed
+- [Compiler] Other columns in a non-grouped aggregate result set are always nullable
 
 ## [2.3.2] - 2026-03-16
 [2.3.2]: https://github.com/sqldelight/sqldelight/releases/tag/2.3.2